refactor: moving test oas files into the __datasets__ dir
