+++ __tests__/__datasets__/nested-allof-flattening.json
