+++ __tests__/__datasets__/path-variable-quirks.json
